Switch to cells feed
The list feed returns the formatted value (in whichever locale is set for the current spreadsheet) for numbers only! This could mean that numbers were included as 123'456.79 (whereas the underlying numeric value was 123456.789.

The cells feed includes a property `numericValue` on each cell which has numeric content (whereas `$t` still contains the formatted value). This way we don't have to rely on string parsing. As a drawback, the desired output of an array of row objects has to be constructed rather tediously from the individual cells.